+++ eslint-bridge/src/rules/main.ts
@@ -120,6 +120,7 @@ import { rule as sqlQueries } from './sql-queries';
 import { rule as standardInput } from './standard-input';
 import { rule as todoTag } from './todo-tag';
 import { rule as unusedImport } from './unused-import';
+import { rule as updatedConstVar } from './updated-const-var';
 import { rule as updatedLoopCounter } from './updated-loop-counter';
 import { rule as useTypeAlias } from './use-type-alias';
 import { rule as variableName } from './variable-name';
@@ -230,6 +231,7 @@ ruleModules['sql-queries'] = sqlQueries;
 ruleModules['standard-input'] = standardInput;
 ruleModules['todo-tag'] = todoTag;
 ruleModules['unused-import'] = unusedImport;
+ruleModules['updated-const-var'] = updatedConstVar;
 ruleModules['updated-loop-counter'] = updatedLoopCounter;
 ruleModules['use-type-alias'] = useTypeAlias;
 ruleModules['variable-name'] = variableName;
